Added 3 monitoring fields:
Upcoming exams: This is informational and just lets us predict the manpower needed to handle support tickets.

Stuck exports: The exporter has error handling, but in the case that it fails, we may have to null out the last_exported field in the database for users.

Duplicate attachments: We believe browser caching may cause the same cloud URL to be reused. We added cache control headers to compensate.